Catch `AssertionError` thrown by step-fn
Fixes https://github.com/mhjort/clj-gatling/issues/81

`(assert ...)` is a useful form for checking the status of a step, but
`AssertionError` inherits from `Throwable`, not `Exception`. We don't
want to catch _all_ throwables, as that includes things like
`VirtualMachineError`, which indicate a fundamental issue where the only
thing to do is exit.

As such, we special-case `AssertionError` as a type to catch alongside
`Exception`.