Improve simulation startup
This aims to improve the results we've been seeing when trying to run
'heavyweight' simulations with `clj-gatling`. Specifically, we are using
`clojure.spec.alpha` and `clojure.spec.gen.alpha` in scenario pre-hooks
to generate the data we want to use within the tests. We did try
pre-generating all the data, but because of the length of test we wanted
this just meant we ran out of heap space...

At low user/max concurrency numbers, the existing setup works more or
less OK. However at high user numbers, especially when there are a lot
of objects to generate per run, we are seeing a pretty big latency spike
at the test start, which is skewing results.

Initially, we thought this was at the system-under-test end, as often
happens when systems recieve a lot of load suddenly and have to scale to
handle it, but further testing showed it was actually a problem within
`clj-gatling` itself. The problem is also present (albiet less
seriously) even without heavyweight pre-hooks, when max concurrency is
high. This is simply because `clj-gatling` is trying to spin up a load
of new runners, single threaded, while simultaneously handling the first
of the scenario runs via `core.async`.

The first fix for this is very simple - swap the `map` where we create
all the different per-user scenario runs for a `pmap`, so we utilise all
the cores available when running the pre-hooks/creating the scenario
runners.

Then we implement an 'initialisation synchronisation' mechanism, so that
all the runners must be ready to go before any of them start sending
requests. We also generate the context data (call pre-hooks) for those
inital requests before reporting ready. Calling all those pre-hooks in
parallel generates a lot of unhelpful load on the `clj-gatling` host
that affects reported timing, whereas subsequent requests are natually
jittery enough from differences in response times from the system under
test that it's fine to regenerate ad-hoc.

`simulation-start` has been turned into an `atom`, that only gets set
when all runners are ready to go. This avoids a problem that is
particularly bad in the `rate` runner, where the first few runners
started can't keep up with the required rate, so once more _do_ start up
there is a higher-than-intended rush of requests to make up the backlog.
Various progress functions etc have been amended to handle this being an
atom/`nil`.

A bug was fixed where the `rate` runner considered a full scenario to be
a single request (and decided whether to continue running or not based
on that), whereas actual sent requests were counted per-step in each
scenario. To aid this, a function has been created which identifies the
number of requests a scenario is expected to make. Most of the time this
is known exactly. In the cases where it is not, a best effort guess is
made, which should get more accurate as more simulation runs occur.

In terms of requests, a simulation is now considered good to run if the
total of sent+pending requests _before_ this run's expected count is
added is less than the max. To aid that, the sent+pending atoms have
been collapsed into a single atom, so that both values can be read...
atomically. Adding this logic into the concurrent runner too has
improved test reliability.

The testing `is-approximately-sorted?` function has been improved, which
allows us to compare without dropping data, and to provide an accuracy
that we want to hit as a target.

Target/actual scenario runs tend to diverge a lot more than expected
when the total test duration (either directly, or from a low max request
count) is very low, so some tests have had the expected number of
requests increased. This has not noticeably affected the overall test
suite duration.

Also stopped the progress reporter from looking odd when the line length
shortens between refreshes, and the sent-messages count looks masssively
too large.