Fix `splits-rate-to-weighted-scenarios` edge case
Fixes #84

Decided to have a look at this as it was one of the tests that failed
and prevented 0.17.6 from being released.

If a very small number of users is specified, close/equal to the number
of scenarios, and the scenario weights are very diverse, you can get a
situation where more users than exist are assigned across the scenarios
(because each must have at least 1 user, you can get a negative
`mismatch` in `split-by-weight`).

Rather than try all sorts of complex maths to deal with this, I opted to
just throw an explicit error detailing the issue in this (edge) case,
and modify the generated tests so that they shouldn't hit it (as well as
creating an expicit test _for_ it).

Hopefully this will mean 0.17.6 can actually get released.