Allow a little flexibility in the force-stop test
Sometimes, because we now start more threads than concurrency calls for
and each checks to see if it should run or not, we can be a little out
in some of the tests. We can't use `approximately==` here, because the
_actualy_ number we expect is normally pretty close to 1, but might be 2
or 3, and we don't want a test that allows a multi-hundred-percent
difference if that number for some reason jumps to 100 or something.

As such, we simply check that the result is within a range from the
expected number to 1 or 2 higher (it shouldn't ever be lower).