Fix a bug of using incorrect concurrency with multiple scenarios
PR87 contained a lot of performance optimisations to clj-gatling itself.
https://github.com/mhjort/clj-gatling/pull/87

Unfortunately it also introduced a bug where end concurrency was incorrect
with multiple scenarios. When using rate mode it makes sense to create bunch
of spare users. However, with concurrency mode we should not have more
users than the total concurrency is.

This commit also adds a test where multiscenario simulation is run and
progress tracker is used for checking that actual concurrency in simulation
is correct.